refactored printed itinerary code
